Implementation and Bug Fixing
Implements Recursive Transformation matrix for children of graph scene.
Put the ior variable in the MaterialPBR, and the PBR fragment shader.
Remove the dragon in the scene and add the Aquarium structure.
Fix the bug that make every element draw at each frame,
Fix the Bounding box for the furstum culling that was missing the GameObject attachment.